[#115362749] fix rounding

+++ ambassadorsdk/src/main/java/com/ambassador/ambassadorsdk/ConversionParameters.java
@@ -71,7 +71,7 @@ public final class ConversionParameters {
     }
 
     public float getRevenue() {
-        return (float) (((int) (revenue * 100)) / 100);
+        return ((int) (revenue * 100)) / 100f;
     }
 
     public int getDeactivateNewAmbassador() {
